Define the first scenario "Search repositories by keyword"
We start by writing a unit test for your scenario.
This Java object will represent a single flow of your application and will encapsulate the business rules behind that user story.
Going test first allow us to discover the various dependencies that our scenario needs.
This package structure clearly reflects the intent of the application: all supported user stories sit on a package, and the boundaries should be as decoupled as possible from the rest of the application.