Add an end-to-end test for the first scenario
The scenario we created will interact with plain java object, which represent business concepts.
In order to actually start implementing the scenario in our application we have to deal with android specific concepts, so our IT test will remain red until we have implemented an android-specific presentation layer.
We decide to use Robotium as our IT library as it make easier to write expressive tests.
For now we're running our end to end tests over the network, but we will use a mock api boundary as soon as we implement one.