Parse a search repository response into Java models
We decide to use Jackson as a Json parsing library for its performance, flexibility and annotation-based configuration.
To ensure maximum flexibility, we hide the ObjectMapper object with our own implementation, JsonConverter.
As you see, we use custom converters for every entity, so that we create our entities through the constructor.
This way our models remain immutable and we can enforce class invariants such as "startDate < endDate".
Also, this keeps Jackson as an implementation detail and we're free to swap for another library (e.g, Gson) without touching our entities.
It is worth noting that entities that are Collections should be first class objects, because it is represented by a concept and this way we can add custom behaviour to it if necessary.